% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{reshape dataset}
\usage{
clean_data(data, value_name)
}
\arguments{
\item{data}{A dataframe containing columns 'iso', 'X2000' to 'X2019'.}

\item{value_name}{A string specifying the name for the value column in the reshaped dataframe.}
}
\value{
A dataframe in long format with columns 'iso', 'year', and the specified value column.
}
\description{
This function reshapes a dataset from wide to long format.
}
\examples{
data("mat_data")
clean_data(mat_data, matmor)
library(dplyr)
library(tidyr)
}
\author{
Xinyi Zhao
}
